+++ src/main/java/subway/line/controller/LineController.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.controller;
 
 import static org.springframework.http.ResponseEntity.status;
 
@@ -15,6 +15,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import subway.line.view.LineCreateRequest;
+import subway.line.view.LineModifyRequest;
+import subway.line.exception.LineNotFoundException;
+import subway.line.view.LineResponse;
+import subway.line.service.LineService;
 
 @RestController
 @RequiredArgsConstructor
