+++ src/main/java/subway/line/exception/LineNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.exception;
 
 public class LineNotFoundException extends RuntimeException {
