+++ src/main/java/subway/line/repository/LineRepository.java
@@ -1,6 +1,8 @@
-package subway.line;
+package subway.line.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import subway.line.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 }
