+++ src/main/java/subway/line/service/LineService.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.service;
 
 import java.util.List;
 import java.util.Optional;
@@ -9,6 +9,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import subway.line.exception.LineNotFoundException;
+import subway.line.repository.LineRepository;
+import subway.line.domain.Line;
+import subway.line.view.LineCreateRequest;
+import subway.line.view.LineModifyRequest;
+import subway.line.view.LineResponse;
 import subway.station.Station;
 import subway.station.StationNotFoundException;
 import subway.station.StationService;
