+++ src/main/java/subway/line/view/LineCreateRequest.java
@@ -1,4 +1,5 @@
-package subway.line;
+package subway.line.view;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +8,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LineModifyRequest {
+public class LineCreateRequest {
     private String name;
     private String color;
+    private long upStationId;
+    private long downStationId;
+    private int distance;
 }
