+++ src/main/java/subway/line/view/LineModifyRequest.java
@@ -1,5 +1,4 @@
-package subway.line;
-
+package subway.line.view;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,10 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LineCreateRequest {
+public class LineModifyRequest {
     private String name;
     private String color;
-    private long upStationId;
-    private long downStationId;
-    private int distance;
 }
